Split arch-dep and arch-indep parts (PARTIAL WORK!!!)
Some parts of kern-ucore including drivers, debug monitor and a few arch-dependent headers are moved to arch/x86_64.
Use a macro TEST_PAGE instead of hard-code for mm checking.
Makefiles are also modified so that this commit can be compiled and run normally (of course with proper configuration such as TARGET_CC_PREFIX).