Deal with different __builtin_va_list.
The magic type __builtin_va_list may be an array of something or a pointer of something, which makes sharing printfmt.c among all architectures buggy (note that for the first argument of getint in printfmt.c, the original version in googlecode project ucore is a pointer to va_list while we have va_list directly now). Two macros are used instead of the original functions getint and getuint, and it seems that this version can work on both i386 and x64 without any problem.