Simple hello world complete.
(1) Change the way of go's memory allocation: SysReserve now has no
effect;
(2) In ucore, argc and argv are passed through edi and rsi, which
differs from that in unix.  We have to manually push them in rt0;
(3) We no longer put 6.out in _initial since it painfully gains the size
of the repository;
(4) fs and gs are stored/restored during trap, and sys_prctl() would
change the trapframe in order to set fs/gs correctly.